just set up product routes
