+++ server/api/index.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const {isAdmin} = require('./authMiddleware')
+// const {isAdmin} = require('./authMiddleware')
 
 module.exports = router
 
@@ -8,7 +8,8 @@ router.use('/products', require('./products'))
 router.use('/categories', require('./categories'))
 router.use('/productcategories', require('./productCategories'))
 router.use('/carts/', require('./carts'))
-router.use('/admin', isAdmin, require('./admin'))
+// router.use('/admin', isAdmin, require('./admin'))
+router.use('/me', require('./me'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
