+++ server/auth/index.js
@@ -19,29 +19,6 @@ router.post('/login', async (req, res, next) => {
   }
 })
 
-// router.post('/update', async (req, res, next) => {
-//   const {firstName, lastName, address} = req.body
-//   const user = req.user
-
-//   console.log('REQ.USER', req.user)
-
-//   try {
-//     let user = await User.findOne({where: {email: req.body.email}})
-
-//     if (!user) {
-//       console.log('No such user found:', req.body.email)
-//     } else if (!user.correctPassword(req.body.password)) {
-//       console.log('Incorrect password for user:', req.body.email)
-//       res.status(401).send('Wrong username and/or password')
-//     } else {
-//       await user.update({firstName, lastName, address})
-//       res.json(user)
-//     }
-//   } catch (err) {
-//     next(err)
-//   }
-// })
-
 router.post('/signup', async (req, res, next) => {
   try {
     const user = await User.create(req.body)
